third_party/blink/web_tests/external/wpt/css/css-values/if-style-query-expected.txt

This is a testharness.js-based test.
Found 7 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] CSS Values and Units Test: CSS inline if() function 15
  assert_equals: "if(style(--x: initial): true_value; else: false_value)" should be substituted to "true_value". expected "true_value" but got "false_value"
[FAIL] CSS Values and Units Test: CSS inline if() function 16
  assert_equals: "if(style(--x: initial): true_value; else: false_value)" should be substituted to "true_value". expected "true_value" but got "false_value"
[FAIL] CSS Values and Units Test: CSS inline if() function 18
  assert_equals: "if(style(--inherited: inherit): true_value;\n          else: false_value)" should be substituted to "true_value". expected "true_value" but got "false_value"
[FAIL] CSS Values and Units Test: CSS inline if() function 19
  assert_equals: "if(style(--inherited: inherit): true_value;\n          else: false_value)" should be substituted to "true_value". expected "true_value" but got "false_value"
[FAIL] CSS Values and Units Test: CSS inline if() function 49
  assert_equals: "if(style(--number: calc(1 + 2)): true_value;\n          else: false_value)" should be substituted to "true_value". expected "true_value" but got "false_value"
[FAIL] CSS Values and Units Test: CSS inline if() function 53
  assert_equals: "if(style(--color: green): true_value;\n          else: false_value)" should be substituted to "true_value". expected "true_value" but got "false_value"
[FAIL] CSS Values and Units Test: CSS inline if() function 55
  assert_equals: "if(style(--color: green): true_value;\n          else: false_value)" should be substituted to "true_value". expected "true_value" but got "false_value"
Harness: the test ran to completion.
